feat: Aztec.js API for registering a contract class (#4469)
Adds a new registerContractClass method in aztecjs public API, that
takes a wallet and an artifact, and returns the tx request ready to
send.

To make sure this tx can be run, we now also register canonical
contracts in the pxe on startup, so they are always available.

Built on #4467