chore: implement poseidon2 opcode (#4446)
This PR create constraints for poseidon2permutation Noir opcodes.
I added a unit test for the opcode.

It also implements poseidon2 permutation circuit for UltraPlonk. Because
the s-box function would require an increase of the polynomial degree in
Ultraplonk relations, it is not compatible with GoblinPlonk circuits.
So I added a non-optimised UltraPlonk version for Poseidon2 permutation.
This is necessary because Noir requires the UltraPlonk backend and it
would not work if we had only the GoblinPlonk version of the opcode.

Thank you @lucasxia01 for the templatised version of the function.

---------

Co-authored-by: lucasxia01 <lucasxia01@gmail.com>