fix(dsl): Add full recursive verification test (#4658)
This PR adds a test that does a full recursive verification similar to
`TestFullRecursiveComposition` under `recursion_constraint.test.cpp` but
using Noir.

A couple bugs fixes were done to enable the new
`double_verify_nested_proof` test:
- We have an assert in the recursion constraint that checks
`num_public_inputs == RecursionConstraint::AGGREGATION_OBJECT_SIZE` when
generating the dummy transcript for when we do not have valid witness
assignments. If we want to recursively verify a recursive proof that
itself has extra public inputs this assertion makes that impossible (in
fact the cpp test does not use any public inputs for
`TestFullRecursiveComposition` and this may be why this assertion wasn't
hit earlier).
- After moving to a definition of a proof as separate from its public
inputs in (https://github.com/AztecProtocol/aztec-packages/pull/3528) we
were missing the accurate copy constraints in the case of no valid
witness assignment. In the case of a nested proof the public inputs will
contain an aggregation object for which we expect two accurate G1
points. Otherwise, even with the removed assertion above we will always
hit a "trying to invert 0" error when working without valid witness
assignments.