refactor: Remove commitment key copy out of instance (#4893)
This PR removes the duplicated commitment key or PCS verification key in
the prover instance and verifier instance. It instead moves the
commitment keys to the proving/verification keys. As a result, the
recursive verification keys must be updated to actually be used, while
they were previously unused and uninstantiated.