fix: Don't reuse brillig with slice arguments (#5800)
This is a quick fix after
https://github.com/AztecProtocol/aztec-packages/pull/5737 since that PR
assumes that one generated brillig is valid for all calls to the brillig
function. This is however not true, since the generated brillig can
differ if the arguments contains slices. This is because the entry point
codegen depends on the size of the slice.
We currently cannot copy slices of any length into brillig due to the
limitations of
[CALLDATACOPY](https://yp-aztec.netlify.app/docs/public-vm/instruction-set#calldatacopy)
where the size being copied needs to be known at compile-time. I'm going
to chat with the AVM team to see if we can lift this restriction and
make brillig entry points be able to copy in arguments that contain
slices of any length.