feat: Validate globals in public kernel (#6031)
- Adds a new `Globals` to `PublicCircuitPublicInputs`, which just
forwards the globals injected into the public call via the
`PublicContextInputs`.
- Adds a new `Globals` to the kernel's `CombinedConstantData`. This is
zero in private-land, and manually set to whatever globals the sequencer
determines in public-land.
- Validates these globals match between kernel and public call in public
kernels, and that they match the ones in the base rollup circuit.
- Loads block gas fees from the `CombinedConstantData` in the public
kernel to calculate the tx fee.

Fixes https://github.com/AztecProtocol/aztec-packages/issues/5843