feat!: switch `bb` over to read ACIR from nargo artifacts (#6283)
This PR further harmonises the `nargo`/`bb` interface by switching `bb`
over to reading the ACIR bytecode directly out of a `nargo` JSON build
artifact.

I've dropped support for contracts in `nargo info` as we're dropping
this feature anyway and this would require us to write a whole bunch of
extra files which we don't need.

Draft as it builds on #6279 plus requires
https://github.com/AztecProtocol/aztec-packages/issues/6168 to migrate
`nargo info` over.