chore: Pw/refactor bb prover (#6349)
This PR moves all native BB code into a new package used by both the
`pxe` for client circuits and the `prover-client` for server circuits.
This removes some of the duplication.

Some duplication is still present. This will remain until we know more
about the differences in proving client and server circuits.