refactor: note emission (#7003)
Adds a `NoteEmission` struct that is returned when new notes are
created. The struct is the note and a `NoteHashCount` that can be used
when emitting the note to ensure that squashing is possible.

There are fairly large change to return this struct throughout, and at
the same time getting rid of the `broadcast` that we have currently been
using whenever the `create_note` life cycle have been encountered. This
also means that keys are not fed into the lifecycle! Meaning that the
aztec nr libraries are generally getting less contaminated with log
decisions.

Instead of always forcing an emit, the returned value can now be emitted
using the `emit` which can be provided a function. For ease of use, we
have also added `encrypted_note_emission.nr` which have a
`encode_and_encrypt` and `encode_and_encrypt_with_keys` functions.

These functions can be used along with the emit to encrypt the note and
then emit it. This allows us to execute it as:

```rust
    let amount = U128::from_integer(amount);
    storage.balances.sub(from, amount).emit(encode_and_encrypt(&mut context, from, from));
    storage.balances.add(to, amount).emit(encode_and_encrypt(&mut context, from, to));
```

**TODO:**
- [x] Remove `broadcast` from note interface
- [x] Remove `encrypt_and_emit_note` from private context

---

There are some build issues that need to be ironed out, but I think that
the flow is nicer than before as it is more clear when things are
emitted, and for cases where we are doing a `get_note` and then
`replace` it is easier for us to not emit the read but only the update
afterwards.